# docs/swagger.yaml
definitions:
...
  model.PmsBrand:
    properties:
      big_pic:
        description: 是否可空:YES 专区大图
        type: string
      brand_story:
        description: 是否可空:YES 品牌故事
        type: string
      factory_status:
        description: 是否可空:YES 是否为品牌制造商：0->不是；1->是
        type: integer
      first_letter:
        description: 是否可空:YES 首字母
        type: string
      id:
        description: 是否可空:NO
        type: integer
      logo:
        description: 是否可空:YES 品牌logo
        type: string
      name:
        description: 是否可空:YES
        type: string
      product_comment_count:
        description: 是否可空:YES 产品评论数量
        type: integer
      product_count:
        description: 是否可空:YES 产品数量
        type: integer
      show_status:
        description: 是否可空:YES
        type: integer
      sort:
        description: 是否可空:YES
        type: integer
    type: object
  response.ResponseMsg:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: mall
  termsOfService: https://github.com/18211167516/Go-Gin-Api
  title: mall
  version: "1.0"
paths:
  /brand/{id}:
    get:
      consumes:
      - application/json
      description: 获取一个品牌
      operationId: v1/PmsBrandController/Brand
      parameters:
      - description: brand_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/model.PmsBrand'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 获取一个品牌
      tags:
      - 品牌接口
  /brand/create:
    post:
      consumes:
      - application/json
      description: 创建一个品牌
      operationId: v1/PmsBrandController/Create
      parameters:
      - description: pmsBrand
        in: body
        name: pmsBrand
        required: true
        schema:
          $ref: '#/definitions/model.PmsBrand'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/model.PmsBrand'
              type: object
        "500":
          description: failure
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/model.PmsBrand'
              type: object
      security:
      - JWT: []
      summary: 创建品牌
      tags:
      - 品牌接口
  /brand/delete/{id}:
    get:
      consumes:
      - application/json
      description: 删除品牌
      operationId: v1/PmsBrandController/Delete
      parameters:
      - description: brand_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 删除品牌
      tags:
      - 品牌接口
  /brand/list:
    get:
      consumes:
      - application/json
      description: 获取品牌列表
      operationId: v1/PmsBrandController/ListBrand
      parameters:
      - default: 0
        description: page number
        in: query
        name: pageNum
        type: integer
      - default: 3
        description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/model.PmsBrand'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 获取品牌列表
      tags:
      - 品牌接口
  /brand/update/{id}:
    post:
      consumes:
      - application/json
      description: 更新品牌
      operationId: v1/PmsBrandController/Update
      parameters:
      - description: brand_id
        in: path
        name: id
        required: true
        type: integer
      - description: pmsBrand
        in: body
        name: pmsBrand
        required: true
        schema:
          $ref: '#/definitions/model.PmsBrand'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/model.PmsBrand'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 更新品牌
      tags:
      - 品牌接口
...
  /sso/getAuthCode:
    get:
      consumes:
      - application/json
      description: 获取验证码
      operationId: v1/UmsMemberController/GetAuthCode
      parameters:
      - description: telephone
        in: query
        name: telephone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      summary: 获取验证码
      tags:
      - 用户接口
  /sso/info:
    get:
      consumes:
      - application/json
      description: 获取会员信息
      operationId: v1/UmsMemberController/Info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 获取会员信息
      tags:
      - 用户接口
  /sso/login:
    post:
      consumes:
      - application/json
      description: 登录
      operationId: v1/UmsMemberController/Login
      parameters:
      - description: username
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      summary: 登录
      tags:
      - 用户接口
  /sso/refreshToken:
    get:
      consumes:
      - application/json
      description: 刷新token
      operationId: v1/UmsMemberController/RefreshToken
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 刷新token
      tags:
      - 用户接口
  /sso/register:
    post:
      consumes:
      - application/json
      description: 注册
      operationId: v1/UmsMemberController/Register
      parameters:
      - description: username
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      - description: telephone
        in: query
        name: telephone
        required: true
        type: string
      - description: authCode
        in: query
        name: authCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      summary: 注册
      tags:
      - 用户接口
  /sso/updatePassword:
    post:
      consumes:
      - application/json
      description: 会员修改密码
      operationId: v1/UmsMemberController/UpdatePassword
      parameters:
      - description: telephone
        in: query
        name: telephone
        required: true
        type: string
      - description: authCode
        in: query
        name: authCode
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 会员修改密码
      tags:
      - 用户接口
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"